# config/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Watson wherespace",
        "description": "",
        "version": "1.0.0",
        "contact":{
        "name": "OpenNTF",
        "url": "http://www.openntf.org",
        "email": "info@openntf.org"
        }
    },
    "host": "api.rabobank.n",
    "schemes": [
        "https"
    ],
    "basePath": "/v2",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/location/{locationId}": {
            "get": {
                "summary": "Get the location by his id",
                "description": "Return the Location by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "locationId",
                        "in": "path",
                        "description": "The ID of the location.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Location"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Location"
                        }
                    },
                    "400": {
                        "description": "Invalid location ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Location not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create the location by his id",
                "description": "Create the Location by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "locationId",
                        "in": "path",
                        "description": "The ID of the location.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Location"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Location"
                        }
                    },
                    "400": {
                        "description": "Invalid location ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Location not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update the location by his id",
                "description": "Update the Location by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "locationId",
                        "in": "path",
                        "description": "The ID of the location.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Location"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Location"
                        }
                    },
                    "400": {
                        "description": "Invalid location ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Location not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete the location by his id",
                "description": "Delete the Location by his ID",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "locationId",
                        "in": "path",
                        "description": "The ID of the location.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Location"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    },
                    "400": {
                        "description": "Invalid location ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Location not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/location/all":{
            "get":{
                "summary": "Get all locations by person",
                "description": "Get the Locations",
                "produces": [
                    "application/json"
                ],

                "tags": [
                    "Location"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Location"
                            }
                        }
                    },
                    "404": {
                        "description": "Location not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/people/event/{eventId}":{
            "get":{
                "summary": "Get all people by event",
                "description": "Get all people by Event",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "The ID of the event.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "People"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Person"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid event ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "People not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/people/location/{locationId}":{
            "get":{
                "summary": "Get all people by location",
                "description": "Get all people by Location",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "locationId",
                        "in": "path",
                        "description": "The ID of the location.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "People"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Person"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Location ID supplied",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "People not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Person": {
            "type": "object",
            "properties": {
                "personId": {
                    "type": "string"
                },
                "emailaddress": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "office": {
                    "type": "string"
                }
            }
        },
        "Location": {
            "type": "object",
            "properties": {
                "locationId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        },
        "Success": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        }
    }
}